requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pillow==10.2.0
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0
django-ckeditor==6.7.0
psycopg2-binary==2.9.9
django-storages[azure]==1.14.2
azure-storage-blob==12.19.0 